# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LocumMap Chennai API (MVP)",
    "description": "FastAPI backend for LocumMap Chennai (MVP).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/assignments/{assignment_id}/otp/create": {
      "post": {
        "tags": [
          "otp"
        ],
        "summary": "Create Otp",
        "operationId": "create_otp_assignments__assignment_id__otp_create_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assignment Id"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(checkin|checkout)$",
              "title": "Type"
            }
          },
          {
            "name": "allow_remote",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Clinic override to allow OTP outside geofence",
              "default": false,
              "title": "Allow Remote"
            },
            "description": "Clinic override to allow OTP outside geofence"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpCreateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assignments/{assignment_id}/otp/verify": {
      "post": {
        "tags": [
          "otp"
        ],
        "summary": "Verify Otp",
        "operationId": "verify_otp_assignments__assignment_id__otp_verify_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assignment Id"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(checkin|checkout)$",
              "title": "Type"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/devices/push-token": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Register Push Token",
        "operationId": "register_push_token_devices_push_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPushTokenIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterPushTokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OtpCreateOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "dev_otp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dev Otp"
          }
        },
        "type": "object",
        "required": [
          "expires_at"
        ],
        "title": "OtpCreateOut"
      },
      "OtpVerifyIn": {
        "properties": {
          "otp": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Otp"
          }
        },
        "type": "object",
        "required": [
          "otp"
        ],
        "title": "OtpVerifyIn"
      },
...
      "RegisterPushTokenIn": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 255,
            "minLength": 10,
            "title": "Token"
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "default": "expo"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "RegisterPushTokenIn"
      },
      "RegisterPushTokenOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "title": "RegisterPushTokenOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}